Move authentication into a separate class
