+++ smart-library-app.xcodeproj/xcuserdata/rileykeane.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -28,7 +28,7 @@
 		<key>smart-library-app.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>1</integer>
+			<integer>0</integer>
 		</dict>
 	</dict>
 </dict>
