+++ smart-library-app/Service/AuthController.swift
@@ -1,0 +1,95 @@
+//
+//  AuthController.swift
+//  smart-library-app
+//
+//  Created by Riley Keane on 10/17/22.
+//
+
+import Foundation
+import CryptoKit
+import FirebaseAuth
+import AuthenticationServices
+
+class AuthController {
+    private var currentNonce: String?
+    private var authResult: AuthDataResult?
+    
+    func requestSignInWithApple(request: ASAuthorizationAppleIDRequest) {
+        let nonce = randomNonceString()
+        currentNonce = nonce
+        request.requestedScopes = [.fullName, .email]
+        request.nonce = sha256(nonce)
+    }
+    
+    func completeSignInWithApple(result: Result<ASAuthorization, Error>) async throws -> AuthDataResult? {
+        switch result {
+        case .success(let authResults):
+            
+            guard let appleIDCredential = authResults.credential as? ASAuthorizationAppleIDCredential else {
+              return nil
+            }
+            
+            guard let nonce = currentNonce else {
+                fatalError("Invalid state: A login callback was received, but no login request was sent.")
+            }
+            
+            guard let appleIDToken = appleIDCredential.identityToken else {
+                fatalError("Invalid state: A login callback was received, but no login request was sent.")
+            }
+            guard let idTokenString = String(data: appleIDToken, encoding: .utf8) else {
+                print("Unable to serialize token string from data: \(appleIDToken.debugDescription)")
+                return nil
+            }
+                
+            let credential = OAuthProvider.credential(withProviderID: "apple.com",idToken: idTokenString,rawNonce: nonce)
+            self.authResult = try await Auth.auth().signIn(with: credential)
+        default:
+            break
+        }
+        return authResult
+    }
+    
+    // from https://firebase.google.com/docs/auth/ios/apple
+    private func randomNonceString(length: Int = 32) -> String {
+      precondition(length > 0)
+      let charset: Array<Character> =
+          Array("0123456789ABCDEFGHIJKLMNOPQRSTUVXYZabcdefghijklmnopqrstuvwxyz-._")
+      var result = ""
+      var remainingLength = length
+
+      while remainingLength > 0 {
+        let randoms: [UInt8] = (0 ..< 16).map { _ in
+          var random: UInt8 = 0
+          let errorCode = SecRandomCopyBytes(kSecRandomDefault, 1, &random)
+          if errorCode != errSecSuccess {
+            fatalError("Unable to generate nonce. SecRandomCopyBytes failed with OSStatus \(errorCode)")
+          }
+          return random
+        }
+
+        randoms.forEach { random in
+          if remainingLength == 0 {
+            return
+          }
+
+          if random < charset.count {
+            result.append(charset[Int(random)])
+            remainingLength -= 1
+          }
+        }
+      }
+
+      return result
+    }
+    
+    //Hashing function using CryptoKit
+    private func sha256(_ input: String) -> String {
+        let inputData = Data(input.utf8)
+        let hashedData = SHA256.hash(data: inputData)
+        let hashString = hashedData.compactMap {
+        return String(format: "%02x", $0)
+        }.joined()
+
+        return hashString
+    }
+}
